bugfix in client and added globals to server
Fixed a client-side bug causing the server to enter an infinite loop
when a client-side abend occurs. This was the result of the client not
sending the proper "exit" directive to the server on abend. A catchall
try/except statement was added to the get_function routine, sending the
exit directive on all unhandled exceptions. This includes the
termination of the client by keyboard interrupt.

Also, a few additional global config options were added to the server in
anticipation of the implementation of AES encryption and a password
file.